chore(all): restructure
