+++ .gitignore
@@ -1,2 +1,3 @@
 *.mp4
 *.log
+app
