+++ worker.go
@@ -9,43 +9,13 @@ import (
 	"github.com/google/uuid"
 )
 
-type ProcessContext struct {
-	WorkerId  int
-	ProcessId string
-	Params    any
-	Stdin     *os.File
-	Stderr    *os.File
-	Stdout    *os.File
-}
-
-func (ctx *ProcessContext) Logger(message any) {
-	content := fmt.Sprintf("[WORKER %d]: %v\n", ctx.WorkerId, message)
-	// fmt.Printf(content)
-
-	_ = ctx.WriteToLogFile(content)
-}
-
-func (ctx *ProcessContext) WriteToLogFile(content string) error {
-	filepath := filepath.Join("logs", fmt.Sprintf("%s.log", ctx.ProcessId))
-
-	file, err := os.OpenFile(filepath, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
-	if err != nil {
-		fmt.Printf("Error opening log file: %v\n", err.Error())
-		return err
-	}
-	defer file.Close()
-
-	file.WriteString(content)
-
-	return nil
-}
-
 type Task interface {
 	Process(taskContext *ProcessContext) error
 }
 
 type WorkerPool struct {
 	Concurreny int
+	LogPath    string
 	tasksChan  chan Task
 	wg         sync.WaitGroup
 	doneChan   chan struct{}
@@ -77,10 +47,14 @@ func (wp *WorkerPool) Wait() {
 
 func (wp *WorkerPool) worker(workerId int) {
 	for task := range wp.tasksChan {
-		fmt.Printf("Worker started %d\n", workerId)
+		if wp.LogPath == "" {
+			wp.LogPath = "logs"
+		}
+
 		ctx := &ProcessContext{
 			WorkerId:  workerId,
 			ProcessId: uuid.NewString(),
+			LogPath:   wp.LogPath,
 		}
 
 		filepath := filepath.Join("logs", fmt.Sprintf("%s.log", ctx.ProcessId))
@@ -107,21 +81,4 @@ func (wp *WorkerPool) worker(workerId int) {
 		wp.wg.Done()
 		fmt.Printf("Worker %d finished\n", workerId)
 	}
-
-}
-
-func (wp *WorkerPool) Run() {
-	wp.tasksChan = make(chan Task)
-
-	for i := 0; i < wp.Concurreny; i++ {
-		go wp.worker(i)
-	}
-
-	// wp.wg.Add(len(wp.Tasks))
-	// for _, task := range wp.Tasks {
-	// 	wp.tasksChan <- task
-	// }
-	close(wp.tasksChan)
-
-	wp.wg.Wait()
 }
